MakeIt/requirements.txt
asgiref==3.7.2
build==0.9.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==3.4.6
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-cloud-videointelligence==2.13.3
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
inflection==0.5.1
multipart==0.2.4
packaging==24.0
pdfservices-sdk==2.3.1
pep517==0.13.0
polling==0.3.2
polling2==0.5.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
uritemplate==4.1.1
urllib3==2.2.1

cloudinary~=1.39.1